fix: Zigbee Switch — cascade control, master enable, and safety fixes
Seven critical bugs prevented Zigbee Switch devices (IKEA E1743 etc.)
from working analogously to physical Switch pins (topin==3):

1. Parent slot was not configured during Switch learning — the head slot
   (zbi, ep=0) never received ZBEE_ROLE_SWITCH, so it was invisible in
   UI and could not trigger PinsLinks. Now the head slot is configured
   with ep=0, ZBEE_ROLE_SWITCH, and VBTN_MODE_PASSTHROUGH after all
   child slots are created.

2. processPins() used uint8_t for the pin ID parameter — values >= 256
   (NUMPIN + high slot index) would overflow. Changed to uint16_t in
   zagotovka.c, zagotovka.h, and main.h.

3. OutputTask sent ZCL commands to Switch slots (which are triggers, not
   actuators). Added a guard: if zbee_role == ZBEE_ROLE_SWITCH the
   output is skipped with a log message.

4. mqtt_zigbee_trigger_handler did not check the master on/off flag for
   Switch slots — the switch worked even when disabled in UI. Added
   onoff==0 check with logging.

5. action_handler did not cascade through Zigbee Switch — Cron, Sunrise,
   DHT22, etc. could not control devices linked via Switch. Added a new
   branch that calls processPins() for ZBEE_ROLE_SWITCH, matching the
   physical switch (topin==3) behavior.

6. switch_payload_on/off were assigned to child slots instead of the
   parent. Moved the assignment to the head slot (zbi) where they belong.

7. gen_switch_json() did not include the "state" field for Zigbee Switch
   entries — the frontend could not display the current ON/OFF position.
   Added "state":%d to the JSON output.

+++ Core/Inc/main.h
@@ -45,7 +45,7 @@ extern struct mg_connection * volatile s_conn;
 
 extern struct Button button[NUMPIN];
 // Прототип функции processPins
-void processPins(uint8_t i, uint8_t action);
+void processPins(uint16_t i, uint8_t action);
 
 // для sim800l
 extern UART_HandleTypeDef huart2;
